SPARQL endpoint integration module split up
Separated the various parts of the SPARQL endpoint integration into separate modules:
* `sparql/endpoint/ktor/core` contains general types, accessible for all target platforms
* `sparql/endpoint/ktor/client` contains various methods of executing and updating RDF datasets through sparql endpoints for all target platforms
* `sparql/endpoint/ktor/server` contains an endpoint implementation, executing queries received through parameters and returning the results on the JVM
* `sparql/endpoint/cli` contains the standalone server implementation of the endpoint, configurable through a CLI on the JVM

The endpoint implementation now properly chunks returned data streams, using the byte writer output stream
