+++ settings.gradle.kts
@@ -37,7 +37,10 @@ include("sparql:common")
 include("sparql:debugging")
 include("sparql:compiler")
 include("sparql:runtime")
-include("sparql:endpoint")
+include("sparql:endpoint:ktor:core")
+include("sparql:endpoint:ktor:client")
+include("sparql:endpoint:ktor:server")
+include("sparql:endpoint:cli")
 include("sparql:test")
 include("sparql:validator")
