+++ sparql/endpoint/cli/src/jvmMain/kotlin/dev/tesserakt/sparql/endpoint/server/EndpointConfig.kt
@@ -1,0 +1,6 @@
+package dev.tesserakt.sparql.endpoint.server
+
+interface EndpointConfig {
+    val port: Int
+    val slug: String
+}
