+++ sparql/endpoint/cli/src/jvmMain/kotlin/dev/tesserakt/sparql/endpoint/server/Server.kt
@@ -1,0 +1,16 @@
+package dev.tesserakt.sparql.endpoint.server
+
+import io.ktor.server.engine.*
+import io.ktor.server.netty.*
+
+class Server(config: EndpointConfig) {
+
+    private val server = embeddedServer(Netty, port = config.port) {
+        sparqlEndpoint(config.slug)
+    }
+
+    fun run() {
+        server.start(wait = true)
+    }
+
+}
