+++ sparql/endpoint/cli/src/jvmMain/kotlin/dev/tesserakt/sparql/endpoint/server/main.kt
@@ -1,0 +1,8 @@
+package dev.tesserakt.sparql.endpoint.server
+
+import com.github.ajalt.clikt.core.main
+
+fun main(args: Array<String>) =
+    CliEntryPoint(
+        run = { config -> Server(config).run() }
+    ).main(args)
