+++ sparql/endpoint/ktor/client/src/commonMain/kotlin/dev/tesserakt/sparql/endpoint/client/ContentNegotiationConfig.kt
@@ -1,0 +1,9 @@
+package dev.tesserakt.sparql.endpoint.client
+
+import dev.tesserakt.sparql.endpoint.core.SparqlContentType
+import io.ktor.client.plugins.contentnegotiation.*
+
+
+fun ContentNegotiationConfig.sparql() {
+    register(SparqlContentType.JsonBindings, SparqlBindingsContentConverter)
+}
