+++ sparql/endpoint/ktor/client/src/commonMain/kotlin/dev/tesserakt/sparql/endpoint/client/SparqlUpdateRequestBuilder.kt
@@ -1,0 +1,37 @@
+package dev.tesserakt.sparql.endpoint.client
+
+import dev.tesserakt.rdf.trig.serialization.TriGSerializer
+import dev.tesserakt.rdf.types.MutableStore
+import dev.tesserakt.rdf.types.factory.mutableStoreOf
+
+
+class SparqlUpdateRequestBuilder(
+    internal val additions: MutableStore = mutableStoreOf(),
+    internal val deletions: MutableStore = mutableStoreOf(),
+) {
+    fun optimise(): SparqlUpdateRequestBuilder {
+        val common = additions.intersect(deletions)
+        if (common.isEmpty()) {
+            return this
+        }
+        additions.removeAll(common)
+        deletions.removeAll(common)
+        return this
+    }
+
+    fun toQueryString(): String = buildString {
+        if (additions.isNotEmpty()) {
+            append("INSERT DATA { ")
+            TriGSerializer.serialize(additions).forEach { append(it) }
+            append(" }")
+        }
+        if (additions.isNotEmpty() && deletions.isNotEmpty()) {
+            append('\n')
+        }
+        if (deletions.isNotEmpty()) {
+            append("DELETE DATA { ")
+            TriGSerializer.serialize(deletions).forEach { append(it) }
+            append(" }")
+        }
+    }
+}
