+++ sparql/endpoint/ktor/server/src/jvmMain/kotlin/dev/tesserakt/sparql/endpoint/server/SparqlEndpoint.kt
@@ -1,0 +1,12 @@
+package dev.tesserakt.sparql.endpoint.server
+
+import dev.tesserakt.sparql.endpoint.core.data.SelectResponse
+
+
+interface SparqlEndpoint {
+
+    suspend fun onSelectQueryRequest(query: String): Result<SelectResponse>
+
+    suspend fun onUpdateQueryRequest(query: String): Result<Unit>
+
+}
