+++ sparql/endpoint/ktor/server/src/jvmMain/kotlin/dev/tesserakt/sparql/endpoint/server/impl/SelectResponseFactory.kt
@@ -1,0 +1,19 @@
+package dev.tesserakt.sparql.endpoint.server.impl
+
+import dev.tesserakt.sparql.Bindings
+import dev.tesserakt.sparql.Query
+import dev.tesserakt.sparql.endpoint.core.data.SelectResponse
+import dev.tesserakt.sparql.endpoint.core.data.SelectResponse.Results.Companion.encoded
+import dev.tesserakt.sparql.evaluation.DeferredOngoingQueryEvaluation
+import dev.tesserakt.sparql.variables
+
+fun SelectResponse(query: Query<Bindings>, evaluation: DeferredOngoingQueryEvaluation<Bindings>): SelectResponse {
+    return SelectResponse(
+        head = SelectResponse.Head(
+            variables = query.variables
+        ),
+        results = SelectResponse.Results(
+            bindings = evaluation.results.map { it.associate { it.first to it.second.encoded() } }
+        )
+    )
+}
