--- sparql/endpoint/src/jvmMain/kotlin/dev/tesserakt/sparql/endpoint/EndpointConfig.kt
@@ -1,6 +1,0 @@
-package dev.tesserakt.sparql.endpoint
-
-interface EndpointConfig {
-    val port: Int
-    val slug: String
-}
