--- sparql/endpoint/src/jvmMain/kotlin/dev/tesserakt/sparql/endpoint/SparqlEndpoint.kt
@@ -1,12 +1,0 @@
-package dev.tesserakt.sparql.endpoint
-
-import dev.tesserakt.sparql.endpoint.data.SelectResponse
-
-
-interface SparqlEndpoint {
-
-    suspend fun onSelectQueryRequest(query: String): Result<SelectResponse>
-
-    suspend fun onUpdateQueryRequest(query: String): Result<Unit>
-
-}
