--- sparql/endpoint/src/jvmMain/kotlin/dev/tesserakt/sparql/endpoint/main.kt
@@ -1,8 +1,0 @@
-package dev.tesserakt.sparql.endpoint
-
-import com.github.ajalt.clikt.core.main
-
-fun main(args: Array<String>) =
-    CliEntryPoint(
-        run = { config -> Server(config).run() }
-    ).main(args)
